feat: add make targets, refactor ful network and temp network tests in different test suites
