+++ packages/client-sdk-nodejs/Makefile
@@ -1,4 +1,4 @@
-.PHONY: test-auth-service test-cache-service test-leaderboard-service test-storage-service test-topics-service test-http-service
+.PHONY: test-auth-service test-cache-service test-leaderboard-service test-storage-service test-topics-service test-http-service test-full-network-outage-retry-service test-temporary-network-outage-retry-service
 
 
 test-auth-service:
@@ -28,3 +28,11 @@ test-topics-service:
 test-http-service:
 	@echo "Testing http service..."
 	@npm run integration-test-http
+
+test-full-network-outage-retry-service:
+	@echo "Testing full network outage retry service..."
+	@npm run integration-test-retry-full-network-outage
+
+test-temporary-network-outage-retry-service:
+	@echo "Testing temporary network outage retry service..."
+	@npm run integration-test-retry-temporary-network-outage
