+++ packages/client-sdk-nodejs/package.json
@@ -23,6 +23,8 @@
     "integration-test-storage": "jest storage/ --maxWorkers 1 -- useConsistentReads",
     "integration-test-store": "npm run integration-test-storage",
     "integration-test-topics": "jest topics/ webhooks/ --maxWorkers 1 -- useConsistentReads",
+    "integration-test-retry-full-network-outage": "jest cache/automated-retry-full-network-outage.test.ts --maxWorkers 1 -- useConsistentReads",
+    "integration-test-retry-temporary-network-outage": "jest cache/automated-retry-temporary-network-outage.test.ts --maxWorkers 1 -- useConsistentReads",
     "integration-test-consistent-reads": "jest integration --maxWorkers 1 -- useConsistentReads",
     "integration-test": "jest integration --maxWorkers 1",
     "unit-test": "jest unit",
