+++ packages/client-sdk-nodejs/test/momento-rpc-method.ts
@@ -3,14 +3,18 @@ export enum MomentoRPCMethod {
   Set = '_SetRequest',
   Delete = '_DeleteRequest',
   Increment = '_IncrementRequest',
+  SetIf = '_SetIfRequest',
+  SetIfNotExists = '_SetIfNotExistsRequest',
   GetBatch = '_GetBatchRequest',
   SetBatch = '_SetBatchRequest',
   KeysExist = '_KeysExistRequest',
+  UpdateTtl = '_UpdateTtlRequest',
   ItemGetTtl = '_ItemGetTtlRequest',
   ItemGetType = '_ItemGetTypeRequest',
   DictionaryGet = '_DictionaryGetRequest',
   DictionaryFetch = '_DictionaryFetchRequest',
   DictionarySet = '_DictionarySetRequest',
+  DictionaryIncrement = '_DictionaryIncrementRequest',
   DictionaryDelete = '_DictionaryDeleteRequest',
   DictionaryLength = '_DictionaryLengthRequest',
   SetFetch = '_SetFetchRequest',
@@ -19,13 +23,23 @@ export enum MomentoRPCMethod {
   SetDifference = '_SetDifferenceRequest',
   SetContains = '_SetContainsRequest',
   SetLength = '_SetLengthRequest',
+  SetPop = '_SetPopRequest',
+  ListPushFront = '_ListPushFrontRequest',
+  ListPushBack = '_ListPushBackRequest',
+  ListPopFront = '_ListPopFrontRequest',
+  ListPopBack = '_ListPopBackRequest',
+  ListErase = '_ListEraseRequest',
   ListRemove = '_ListRemoveRequest',
   ListFetch = '_ListFetchRequest',
   ListLength = '_ListLengthRequest',
+  ListConcatenateFront = '_ListConcatenateFrontRequest',
+  ListConcatenateBack = '_ListConcatenateBackRequest',
+  ListRetain = '_ListRetainRequest',
   SortedSetPut = '_SortedSetPutRequest',
   SortedSetFetch = '_SortedSetFetchRequest',
   SortedSetGetScore = '_SortedSetGetScoreRequest',
   SortedSetRemove = '_SortedSetRemoveRequest',
+  SortedSetIncrement = '_SortedSetIncrementRequest',
   SortedSetGetRank = '_SortedSetGetRankRequest',
   SortedSetLength = '_SortedSetLengthRequest',
   SortedSetLengthByScore = '_SortedSetLengthByScoreRequest',
